start menu
I have finished the majority parts of the start menu(apart from the ‘About’ page).
The style of the images are not the same at the moment, I will try to find some more images that fit our game.
The codes are working so it is not hard to change the images.